sgwcd: Fix crash hack
This is obviously not the correct way to fix this problem, because it
does not resolve the underlying cause and it tells the cpe that the
session has been re-established when in fact it has not. But it will
keep the core online for a while so I'm trying it.