[pfcp] sendto should always return OGS_OK (#81)
sendto() was crashing the entire SMF when it received an EDESTADDRREQ error. True source of the error is unknown (maybe wireguard hiccupping? looks like the packet was sent out on the wrong wg tunnel) but regardless of the error we should actually just handle gracefully, log error, and move on. PFCP already has its own retrans/recovery mechanisms so loss here is irrelevant.