Worth Keeping: error handling ignores null check
This particular fix looks like a specific oversight that is worth
keeping regardless of other context. While some of these other patches
may have implications I'm unsure of it's pretty obvious in this
particular error case that we should guard for sess being null and that
guard was simply forgotten.